Исправление бага HomeWorkdaoImpl

+++ src/main/java/ru/top/project/dao/HomeWorkDao.java
@@ -1,16 +1,17 @@
 package ru.top.project.dao;
 import ru.top.project.model.HomeWork;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface HomeWorkDao {
     List<HomeWork> getAllHomeWorks();
 
-    HomeWork getHomeWorkById(int id);
+    HomeWork getHomeWorkById(BigInteger id);
 
     HomeWork addHomeWork(HomeWork homeWork);
 
     HomeWork updateHomeWork(HomeWork homeWork);
 
-    void deleteHomeWork(int id);
+    void deleteHomeWork(BigInteger id);
 }
